feat : Valid 어노테이션 및 주석 추가 + 찜 목록 구현부 수정

+++ src/main/java/com/bootakhae/webapp/product/entities/ProductEntity.java
@@ -24,10 +24,10 @@ public class ProductEntity extends BaseEntity {
     private String name;
 
     @Column(name="product_price", nullable = false)
-    private Integer price;
+    private Long price;
 
     @Column(name="product_stock", nullable = false)
-    private Integer stock;
+    private Long stock;
 
     @Column(name="product_producer", nullable = false, length = 30)
     private String producer;
